Refactor out query for stacktraces
